Refactored testing to use TestItem and parallelization
